# be/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  application.UpdatePodInput:
    properties:
      categories:
        items:
          type: string
        maxItems: 5
        type: array
      description:
        maxLength: 2000
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      tags:
        items:
          type: string
        maxItems: 10
        type: array
      visibility:
        allOf:
        - $ref: '#/definitions/domain.Visibility'
        enum:
        - public
        - private
    type: object
  docs.ErrorBody:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        items:
          $ref: '#/definitions/docs.ErrorDetail'
        type: array
      message:
        example: Invalid input data
        type: string
    type: object
  docs.ErrorDetail:
    properties:
      field:
        example: email
        type: string
      message:
        example: Invalid email format
        type: string
    type: object
  docs.ErrorResponse:
    description: Standard error response format
    properties:
      error:
        $ref: '#/definitions/docs.ErrorBody'
      meta:
        $ref: '#/definitions/docs.Meta'
      success:
        example: false
        type: boolean
    type: object
  docs.MessageResponse:
    description: Success response with a message
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      meta:
        $ref: '#/definitions/docs.Meta'
      success:
        example: true
        type: boolean
    type: object
  docs.Meta:
    properties:
      request_id:
        example: req_abc123
        type: string
      timestamp:
        example: "2025-12-08T10:00:00Z"
        type: string
    type: object
  docs.SuccessResponse:
    description: Standard success response format
    properties:
      data: {}
      meta:
        $ref: '#/definitions/docs.Meta'
      success:
        example: true
        type: boolean
    type: object
  docs.SuggestedQuestionsResponse:
    description: Suggested questions for AI chat
    properties:
      questions:
        example:
        - What are the key concepts?
        - Can you explain the main idea?
        - What are the practical applications?
        items:
          type: string
        type: array
    type: object
...
  domain.Pod:
    properties:
      categories:
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      fork_count:
        type: integer
      forked_from_id:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      slug:
        type: string
      star_count:
        type: integer
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
      view_count:
        type: integer
      visibility:
        $ref: '#/definitions/domain.Visibility'
    type: object
  domain.Visibility:
    enum:
    - public
    - private
    type: string
    x-enum-varnames:
    - VisibilityPublic
    - VisibilityPrivate
  errors.ErrorBody:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/errors.ErrorDetail'
        type: array
      message:
        type: string
    type: object
  errors.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
      value: {}
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorBody'
      meta:
        $ref: '#/definitions/errors.ResponseMeta'
      success:
        type: boolean
    type: object
  errors.ResponseMeta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  fiber.Map:
    additionalProperties: true
    type: object
...
  http.ChatRequest:
    properties:
      message:
        maxLength: 2000
        minLength: 1
        type: string
    required:
    - message
    type: object
...
  http.ExportChatRequest:
    properties:
      format:
        enum:
        - pdf
        - markdown
        type: string
    required:
    - format
    type: object
  http.FeedbackRequest:
    properties:
      feedback:
        enum:
        - thumbs_up
        - thumbs_down
        type: string
      feedback_text:
        type: string
    required:
    - feedback
    type: object
...
  http.UpdateMaterialRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      title:
        maxLength: 255
        type: string
    type: object
...
  response.Meta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
...
  response.Response-domain_Pod:
    properties:
      data:
        $ref: '#/definitions/domain.Pod'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@ngasihtau.com
    name: NgasihTau API Support
  description: |-
    NgasihTau is a learning platform API that enables teachers to create Knowledge Pods for sharing learning materials with students.
    The platform uses a microservices architecture with Go as the primary backend language.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: NgasihTau API
  version: "1.0"
paths:
...
  /chat/{messageId}/feedback:
    post:
      consumes:
      - application/json
      description: Submit thumbs up/down feedback for an AI response
      parameters:
      - description: Message ID
        format: uuid
        in: path
        name: messageId
        required: true
        type: string
      - description: Feedback data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.FeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Feedback submitted
          schema:
            $ref: '#/definitions/docs.MessageResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit feedback
      tags:
      - AI
...
  /materials/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-delete a material
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Material deleted
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Delete a material
      tags:
      - Materials
    get:
      consumes:
      - application/json
      description: Get a material by ID
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Material details
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Material not found
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Get a material
      tags:
      - Materials
    put:
      consumes:
      - application/json
      description: Update a material's title and description
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Material update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated material
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Update a material
      tags:
      - Materials
...
  /materials/{id}/chat:
    post:
      consumes:
      - application/json
      description: Send a message to AI with material context using RAG
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Chat message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: AI response
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Chat with material
      tags:
      - AI
  /materials/{id}/chat/export:
    post:
      consumes:
      - application/json
      description: Export chat history for a material to PDF or Markdown format
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Export format
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ExportChatRequest'
      produces:
      - application/pdf
      - text/markdown
      responses:
        "200":
          description: Exported chat file
          schema:
            type: file
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export chat history
      tags:
      - AI
  /materials/{id}/chat/history:
    get:
      consumes:
      - application/json
      description: Get paginated chat history for a material
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: Maximum messages
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Chat history
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get chat history
      tags:
      - AI
  /materials/{id}/chat/suggestions:
    get:
      consumes:
      - application/json
      description: Get AI-generated suggested questions based on material content
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Suggested questions
          schema:
            $ref: '#/definitions/docs.SuggestedQuestionsResponse'
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get suggested questions
      tags:
      - AI
...
  /pods/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Knowledge Pod. Only the owner can delete a pod.
      parameters:
      - description: Pod ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Pod deleted successfully
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Owner access required
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Pod not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a Knowledge Pod
      tags:
      - Pods
    get:
      consumes:
      - application/json
      description: Get a Knowledge Pod by ID. Private pods require authentication
        and access.
      parameters:
      - description: Pod ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pod details
          schema:
            $ref: '#/definitions/response.Response-domain_Pod'
        "400":
          description: Invalid pod ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Access denied to private pod
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Pod not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a Knowledge Pod
      tags:
      - Pods
    put:
      consumes:
      - application/json
      description: Update a Knowledge Pod. Requires owner or admin collaborator access.
      parameters:
      - description: Pod ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Pod update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application.UpdatePodInput'
      produces:
      - application/json
      responses:
        "200":
          description: Updated pod
          schema:
            $ref: '#/definitions/response.Response-domain_Pod'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Edit access required
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Pod not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a Knowledge Pod
      tags:
      - Pods
...
  /pods/{id}/chat:
    post:
      consumes:
      - application/json
      description: Send a message to AI with context from all materials in a pod
      parameters:
      - description: Pod ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Chat message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: AI response with material citations
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Chat with pod
      tags:
      - AI
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"